+++ Patches/StormyWeatherPatch.cs
@@ -5,6 +5,25 @@ namespace GeneralImprovements.Patches
 {
     internal static class StormyWeatherPatch
     {
+        public static StormyWeather Instance { get; private set; }
+
+        [HarmonyPatch(typeof(StormyWeather), nameof(OnEnable))]
+        [HarmonyPostfix]
+        private static void OnEnable(StormyWeather __instance)
+        {
+            if (Instance != __instance)
+            {
+                Instance = __instance;
+            }
+        }
+
+        [HarmonyPatch(typeof(StormyWeather), nameof(OnDisable))]
+        [HarmonyPostfix]
+        private static void OnDisable()
+        {
+            Instance = null;
+        }
+
         [HarmonyPatch(typeof(StormyWeather), nameof(SetStaticElectricityWarning))]
         [HarmonyPostfix]
         private static void SetStaticElectricityWarning(NetworkObject warningObject)