+++ Utilities/NetworkHelper.cs
@@ -1,0 +1,30 @@
+﻿using GeneralImprovements.Patches;
+using Unity.Netcode;
+
+namespace GeneralImprovements.Utilities
+{
+    internal class NetworkHelper : NetworkBehaviour
+    {
+        public static NetworkHelper Instance { get; private set; }
+
+        private void Start()
+        {
+            Instance = this;
+        }
+
+        [ClientRpc]
+        public void SyncExtraDataOnConnectClientRpc(int totalDays, int totalDeaths, int daysSinceLastDeath, ClientRpcParams clientParams)
+        {
+            if (!IsServer && StartOfRound.Instance != null)
+            {
+                Plugin.MLS.LogInfo("Client received extra data sync RPC.");
+                StartOfRound.Instance.gameStats.daysSpent = totalDays;
+                StartOfRound.Instance.gameStats.deaths = totalDeaths;
+                StartOfRoundPatch.DaysSinceLastDeath = daysSinceLastDeath;
+
+                MonitorsHelper.UpdateTotalDaysMonitors();
+                MonitorsHelper.UpdateDeathMonitors();
+            }
+        }
+    }
+}